User places request and moderators/admin approve implementation

+++ web/src/places/placesScripts.js
@@ -2,15 +2,33 @@ import {ObjectId} from "mongodb";
 import * as fs from "fs";
 
 async function addNewPlace (req, res, db){
-    await db.collection('places').insertOne({place_name: req.body.placeName, place_description: req.body.placeDescription,
-    usersLiked: [], comments: [], image: `http://localhost:3003/uploads/${req.file.filename}`}).then(res => console.log(res));
+    let pathToImage = null;
+    if(req.file){
+        pathToImage = `http://localhost:3003/uploads/${req.file.filename}`;
+    }
+    if(req.signedCookies.moderator || req.signedCookies.admin){
+        await db.collection('places').insertOne({place_name: req.body.placeName, place_description: req.body.placeDescription,
+            usersLiked: [], comments: [], image: pathToImage, approved: true}).then(res => console.log(res));
+    }
+    else if(req.signedCookies.user){
+        await db.collection('places').insertOne({place_name: req.body.placeName, place_description: req.body.placeDescription,
+            usersLiked: [], comments: [], image: pathToImage, approved: false}).then(res => console.log(res));
+    }
     res.end();
 }
 
 async function sendPlaces (req, res, db){
-    const data = await db.collection('places').find({}).toArray().then(res => res).catch(e => console.log(e));
-    res.send(data);
-
+    let searchParam = null;
+    if(req.params.state === 'approved'){
+        searchParam = {approved : true};
+    }
+    if(req.params.state === 'proposed' && (req.signedCookies.moderator || req.signedCookies.admin)){
+        searchParam = {approved : false};
+    }
+    if(searchParam){
+        const data = await db.collection('places').find(searchParam).toArray().then(res => res).catch(e => console.log(e));
+        res.send(data);
+    }
     res.end;
 
 }
@@ -19,16 +37,29 @@ async function removePlace(req, res, db){
 
     const place = await db.collection('places').findOne({_id: ObjectId(req.body.key)}).catch(e => console.log(e));
     if(place){
+        if(place.image){
+            const imagePath = place.image.split('/').pop();
+            fs.unlink('C:/Users/kirik/WebstormProjects/code_voyager/web/uploads/' + imagePath, (err) => {
+                if(err){
+                    console.log(err)
+                }
+            });
+        }
 
-        const imagePath = place.image.split('/').pop();
-        console.log(imagePath);
         await db.collection('places').deleteOne({_id: ObjectId(req.body.key)}).catch(e => console.log(e));
-        fs.unlink('C:/Users/kirik/WebstormProjects/code_voyager/web/uploads/' + imagePath, (err) => {
-            if(err){
-                console.log(err)
-            }
-        });
+
     }
     res.end();
 }
-export {addNewPlace, sendPlaces, removePlace}
+
+
+async function approvePlace(req, res, db){
+    const update = {
+        $set: {
+            approved: req.body.approved
+        }
+    }
+    await db.collection('places').updateOne({_id: ObjectId(req.params.id)}, update);
+    res.end();
+}
+export {addNewPlace, sendPlaces, removePlace, approvePlace}
